refactor: Sidebar components
- rename `ModuleSidebarLink` to `ModuleSidebarItem` to house Spacer & Section Break in the same component

- use `ModuleSidebarItem` recursively to render sidebar links inside section

- Add Menu component for spacers, section breaks in edit mode

- provide `updateSidebarItem` fn to Menu component (grandchild component) from Sidebar (grandparent component)
