+++ desk/src/components/ModuleSidebarItem.vue
@@ -1,0 +1,110 @@
+<template>
+	<Tooltip
+		placement="right"
+		:text="item.label"
+		:hover-delay="0.1"
+		:disabled="!isCollapsed"
+		v-if="type === 'Link'"
+	>
+		<router-link
+			:to="item.route_to"
+			class="flex cursor-pointer items-center gap-2 truncate rounded px-2 py-1 transition duration-300 ease-in-out"
+			:class="[
+				isCollapsed ? 'justify-center' : '',
+				isActive && !isEditing ? 'bg-white shadow-sm' : 'hover:bg-gray-200',
+				isEditing ? 'group/item' : '',
+			]"
+			v-bind="$attrs"
+		>
+			<Icon :name="item.icon || 'folder-normal'" class="h-5 w-5 text-gray-700" />
+			<div v-if="!isCollapsed" class="flex items-center gap-1 truncate text-base text-gray-700">
+				{{ item.label }}
+			</div>
+			<ModuleSidebarItemMenu :item="item" />
+		</router-link>
+	</Tooltip>
+
+	<template v-else-if="type === 'Spacer'">
+		<div
+			v-if="isEditing"
+			class="group/item ml-2 flex min-h-6 items-center justify-center rounded border-dashed border-gray-400 px-2 text-xs uppercase text-gray-600 hover:border"
+		>
+			<ModuleSidebarItemMenu :item="item" />
+		</div>
+		<div v-else class="h-5"></div>
+	</template>
+
+	<div v-else-if="type === 'Section Break' && item.links?.length">
+		<div v-if="isCollapsed" class="mx-2 my-2 h-1 border-b"></div>
+		<div
+			v-else
+			@click="item.opened = !item.opened"
+			class="group/item mt-5 flex cursor-pointer items-center gap-2 px-2"
+			:class="item.opened ? 'mb-3' : ''"
+		>
+			<FeatherIcon
+				:name="item.opened ? 'chevron-down' : 'chevron-right'"
+				class="h-4 w-4 font-semibold text-gray-600"
+			/>
+			<div class="flex items-center gap-1 text-sm uppercase text-gray-700">
+				{{ item.label }}
+			</div>
+			<ModuleSidebarItemMenu :item="item" />
+		</div>
+
+		<nav v-if="item.opened" class="flex flex-col space-y-0.5">
+			<ModuleSidebarItem
+				v-for="link in item.links"
+				type="Link"
+				:key="link.name"
+				:item="link"
+				:isCollapsed="isCollapsed"
+				:isEditing="isEditing"
+				:showEditMenu="false"
+			/>
+		</nav>
+	</div>
+</template>
+
+<script setup>
+import { computed } from "vue"
+import { useRoute, useRouter } from "vue-router"
+import { Tooltip, FeatherIcon } from "frappe-ui"
+import Icon from "@/components/Icon.vue"
+import ModuleSidebarItemMenu from "@/components/ModuleSidebarItemMenu.vue"
+
+const props = defineProps({
+	type: {
+		type: String,
+		required: true,
+	},
+	item: {
+		type: Object,
+		required: true,
+	},
+	isCollapsed: {
+		type: Boolean,
+		required: false,
+		default: false,
+	},
+	isEditing: {
+		type: Boolean,
+		required: false,
+		default: false,
+	},
+	showEditMenu: {
+		type: Boolean,
+		required: false,
+		default: true,
+	},
+})
+
+const router = useRouter()
+const route = useRoute()
+
+const isActive = computed(() => {
+	const linkRoute = router.resolve(props.item.route_to)
+	// Check if the current route is the same as the link route, with optional trailing slash
+	return route.path.match(new RegExp(`^${linkRoute.path}(/|$)`))
+})
+</script>
